[FIX] Dockerfile Odoo 13.0
For some reason Odoo put `sale_coupon` module in the EE edition with
an only French localisation file in it.
https://github.com/odoo/enterprise/tree/13.0/sale_coupon/i18n
Because we are putting CE and EE addons all together in one directory,
this creates an issue.
So let's just skip this EE module.
And if you need it - you are welcome to develop a patch :)